Join outstanding Robodash threads on exit
When using Robodash in short lived environments like a rake task, the
child thread can be killed by the main thread before it finishes
execution.  This means we could miss pings if they occur at the end of
a rake task.

To avoid this we keep track of the spawned threads and make sure to
join any that are still alive on main application exit.